refactor(job_source): extract shared connection utilities
Move ExponentialBackoff and ConnectOutcome out of stratum_v1.rs into a
new job_source/connection.rs module so both the V1 and V2 sources can
share the same implementation without copy-paste drift.

# Conflicts:
#	mujina-miner/src/job_source/stratum_v1.rs
