+++ mujina-miner/src/job_source/stratum_v1.rs
@@ -4,9 +4,7 @@
 //! abstraction. It handles the conversion between Stratum protocol messages and
 //! the internal JobTemplate/Share types used by the scheduler.
 
-use std::collections::hash_map::RandomState;
 use std::future;
-use std::hash::{BuildHasher, Hasher};
 use std::time::Duration;
 
 use anyhow::Result;
@@ -20,6 +18,7 @@ use crate::stratum_v1::{
 use crate::tracing::prelude::*;
 use crate::types::{Difficulty, HashRate, ShareRate};
 
+use super::connection::{ConnectOutcome, ExponentialBackoff};
 use super::{
     Extranonce2Range, GeneralPurposeBits, JobTemplate, MerkleRootKind, MerkleRootTemplate, Share,
     SourceCommand, SourceEvent, VersionTemplate,
@@ -71,67 +70,6 @@ const MAX_RECONNECT_DELAY: Duration = Duration::from_secs(60);
 /// dead pool for over a decade. Two minutes spans several refresh
 /// intervals.
 const MAX_JOB_GAP: Duration = Duration::from_secs(2 * 60);
-
-/// Exponential backoff for reconnection timing.
-///
-/// Starts at `initial` and doubles after each call to `next_delay()`,
-/// capping at `max`. Each returned delay is jittered to [0.5, 1.0] of
-/// the nominal value to avoid thundering-herd reconnections.
-struct ExponentialBackoff {
-    current: Duration,
-    initial: Duration,
-    max: Duration,
-    // Per-process jitter seed. RandomState is seeded from OS randomness
-    // at construction, so different processes produce different jitter
-    // even when reconnecting at the same wall-clock instant. This is
-    // the same approach tokio uses internally for jittered timeouts.
-    jitter_state: RandomState,
-    jitter_step: u64,
-}
-
-impl ExponentialBackoff {
-    fn new(initial: Duration, max: Duration) -> Self {
-        Self {
-            current: initial,
-            initial,
-            max,
-            jitter_state: RandomState::new(),
-            jitter_step: 0,
-        }
-    }
-
-    /// Return the next backoff delay (with jitter) and advance the state.
-    ///
-    /// The nominal delay (1s, 2s, 4s, ...) is scaled by a jitter factor
-    /// in [0.5, 1.0] to spread out reconnection attempts across miners.
-    fn next_delay(&mut self) -> Duration {
-        let nominal = self.current;
-        self.current = (self.current * 2).min(self.max);
-
-        let mut hasher = self.jitter_state.build_hasher();
-        hasher.write_u64(self.jitter_step);
-        self.jitter_step = self.jitter_step.wrapping_add(1);
-        let hash = hasher.finish();
-        let jitter = 0.5 + (hash as f64 / u64::MAX as f64) * 0.5;
-
-        nominal.mul_f64(jitter)
-    }
-
-    /// Reset backoff to the initial delay.
-    fn reset(&mut self) {
-        self.current = self.initial;
-    }
-}
-
-/// Outcome of a single connection attempt.
-enum ConnectOutcome {
-    /// Graceful shutdown requested.
-    Shutdown,
-    /// Connection lost; retry after backoff.
-    Disconnected,
-    /// Unrecoverable error (e.g. auth failure); stop retrying.
-    Fatal(anyhow::Error),
-}
 
 /// Stratum v1 job source.
 ///
